Refactor : 배틀 퇴장 로직 useBattleLeave 훅으로 분리
